Install ProgBar - Fix 3 fml

+++ app.py
@@ -229,18 +229,23 @@ def main():
                 sgui.root['startInstall'].update(disabled=True)
                 sgui.root['installPBAR'].update(current_count=0)
                 _install_prog_file = os.path.join(app_path, "install_progress.txt")
+                _mem_prog = 0
                 while True:
                     if not os.path.isfile(_install_prog_file):
                         continue
                     with open(_install_prog_file, "r") as fr:
-                        _curr_prog_file = int(fr.read())
+                        _curr_prog_file = int(fr.read().replace("\n", "").strip())
+                    if _mem_prog == _curr_prog_file:
+                        continue
+                    print(f" [ DEBUG ] -> _curr_prog_file = {_curr_prog_file} ")
+                    _mem_prog = _curr_prog_file
                     _curr_prog = (_curr_prog_file/_ammount_models) * 100
                     sgui.root['installPBAR'].update(current_count=_curr_prog)
                     if _curr_prog == 100:
                         os.remove(_install_prog_file)
-                        sgui.root.close()
-                        sgui = SGui(get_user_theme())
                         break
+                sgui.root.close()
+                sgui = SGui(get_user_theme())
